requirements.txt
discord.py
openai
nest_asyncio
ffmpeg-python
